Recognize icon type and convert to QIcon accordingly
If we are dealing with a GFileIcon, g_icon_to_string() might return
a path. If that's the case we need to tell QIcon, otherwise we assume
it's an icon name but if the current theme doesn't have it we load
the "unknown" icon.

See also https://developer.gnome.org/gio/2.32/GIcon.html#g-icon-to-string